requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
qdrant-client==1.7.0
cohere==4.37
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
redis==5.0.1
structlog==23.2.0
prometheus-client==0.19.0
asyncio==3.4.3
tenacity==8.2.3
pydantic-settings==2.1.0
python-json-logger==2.0.7